chore(fs): Rename src/visitor to src/interpreter
