+++ test/interpreter/Arithmetic.test.js
@@ -5,7 +5,7 @@ const Stmt = require("../../lib/parser/Statement");
 const { token } = require("../parser/ParserTests");
 const { binary } = require("./InterpreterTests");
 const { Lexeme } = require("../../lib/Lexeme");
-const { Interpreter } = require("../../lib/visitor/Interpreter");
+const { Interpreter } = require("../../lib/interpreter");
 const BrsTypes = require("../../lib/brsTypes");
 
 let interpreter;
