+++ test/interpreter/For.test.js
@@ -2,7 +2,7 @@ const BrsError = require("../../lib/Error");
 const Expr = require("../../lib/parser/Expression");
 const Stmt = require("../../lib/parser/Statement");
 const { Lexeme } = require("../../lib/Lexeme");
-const { Interpreter } = require("../../lib/visitor/Interpreter");
+const { Interpreter } = require("../../lib/interpreter");
 const { Int32 } = require("../../lib/brsTypes/Int32");
 const { identifier } = require("../parser/ParserTests");
